Cart Table added

+++ src/main/java/com/prog2/ecommerce/repository/ProductRepository.java
@@ -9,9 +9,9 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 
-    //Custom query
+    // Custom query
     @Query(value = "select p from Product p where p.name like %:keyword%")
-    List<Product> findByKeyword( String keyword);
+    List<Product> findByKeyword(String keyword);
 }
