+++ src/main/java/com/prog2/ecommerce/repository/UserRepository.java
@@ -1,6 +1,5 @@
 package com.prog2.ecommerce.repository;
 
-
 import com.prog2.ecommerce.model.User;
 
 import java.util.List;
@@ -10,8 +9,9 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
- //Custom query
- @Query(value = "select p from User p where p.email like %:email%")
- List<User> findByEmail( String email);
+public interface UserRepository extends JpaRepository<User, Integer> {
+    // Custom query
+    @Query(value = "select p from User p where p.email like %:email%")
+    List<User> findByEmail(String email);
+
 }
